+++ ModLoaderTests/ModLoaderTest.cs
@@ -98,10 +98,10 @@ namespace Sheepy.Modnix.Tests {
       [TestMethod()] public void ConflictTest () {
          // A conflicts with B, B conflicts with C, D conflicts with A B D.
          // A disables B, B is skipped, D disables A and B and skip itself, leaving C and D
-         var A = new ModEntry( new ModMeta{ Id = "A", Version = Ver( "1.2" ), Conflicts = new AppVer[]{ new AppVer{ Id = "B" }, new AppVer{ Id = "D", Min = Ver( "4.5" ) } } } );
-         var B = new ModEntry( new ModMeta{ Id = "B", Version = Ver( "2.3" ), Conflicts = new AppVer[]{ new AppVer{ Id = "C" } } } );
-         var C = new ModEntry( new ModMeta{ Id = "C", Version = Ver( "4.5" ), Conflicts = new AppVer[]{ new AppVer{ Id = "D", Max = Ver( "2.0" ) } } } );
-         var D = new ModEntry( new ModMeta{ Id = "D", Version = Ver( "3.4" ), Conflicts = new AppVer[]{ new AppVer{ Id = "A" }, new AppVer{ Id = "B" },  new AppVer{ Id = "D" } } } );
+         var A = new ModEntry( new ModMeta{ Id = "A", Version = Ver( "1.2" ), Disables = new AppVer[]{ new AppVer{ Id = "B" }, new AppVer{ Id = "D", Min = Ver( "4.5" ) } } } );
+         var B = new ModEntry( new ModMeta{ Id = "B", Version = Ver( "2.3" ), Disables = new AppVer[]{ new AppVer{ Id = "C" } } } );
+         var C = new ModEntry( new ModMeta{ Id = "C", Version = Ver( "4.5" ), Disables = new AppVer[]{ new AppVer{ Id = "D", Max = Ver( "2.0" ) } } } );
+         var D = new ModEntry( new ModMeta{ Id = "D", Version = Ver( "3.4" ), Disables = new AppVer[]{ new AppVer{ Id = "A" }, new AppVer{ Id = "B" },  new AppVer{ Id = "D" } } } );
 
          var AllMods = ModScanner.AllMods;
          AllMods.Add( A );